requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
transformers==4.35.2
datasets==2.14.6
peft==0.6.2
torch==2.2.1
rich==13.7.0
accelerate==0.24.1
bitsandbytes==0.41.3
numpy==1.24.4
pandas==2.1.3
Flask==2.3.3